project restructure added next folders:common,page,ui
