+++ src/app/components/common/table/index.js
@@ -1,0 +1,7 @@
+import Table from "./table";
+import TableBody from "./tableBody";
+import TableHeader from "./tableHeader";
+
+export default Table;
+
+export { TableBody, TableHeader };
