+++ src/app/components/page/userPage/userPage.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
-import api from "../api";
+import api from "../../../api";
 import PropTypes from "prop-types";
-import QualitiesList from "./qualitiesList";
+import Qualities from "../../ui/qualities";
 
 const UserPage = ({ userId }) => {
     const history = useHistory();
@@ -26,7 +26,7 @@ const UserPage = ({ userId }) => {
         <div>
             <h1>{user.name}</h1>
             <h2>Профессия: {user.profession.name}</h2>
-            <QualitiesList qualities={user.qualities} />
+            <Qualities qualities={user.qualities} />
             <p>completedMeetings: {user.completedMeetings}</p>
             <h2>Rate: {user.rate}</h2>
             <button className="btn btn-secondary" onClick={handleAllUsers}>
