+++ src/app/components/page/usersListPage/index.js
@@ -1,0 +1,4 @@
+import UsersListPage from "./usersListPage";
+
+export default UsersListPage
+
