+++ src/app/components/page/usersListPage/usersListPage.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from "react";
-import Pagination from "./pagination";
-import { paginate } from "../utils/paginate";
+import Pagination from "../../common/pagination";
+import { paginate } from "../../../utils/paginate";
 import PropTypes from "prop-types";
-import api from "../api";
-import GroupList from "./groupList";
-import SearchStatus from "./searchStatus";
-import UsersTable from "./usersTable";
+import api from "../../../api";
+import GroupList from "../../common/groupList";
+import SearchStatus from "../../ui/searchStatus";
+import UsersTable from "../../ui/usersTable";
 import _ from "lodash";
-import TextField from "./textField";
+import TextField from "../../common/form/textField";
 
 const UsersListPage = () => {
     const [currentPage, setCurrentPage] = useState(1);
