+++ src/app/components/ui/qualities/index.js
@@ -1,0 +1,3 @@
+import Qualities from "./qualitiesList";
+
+export default Qualities
