+++ src/app/components/ui/qualities/qualitiesList.jsx
