+++ src/app/components/ui/usersTable.jsx
@@ -1,10 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import TableHeader from "./tableHeader";
-import TableBody from "./tableBody";
-import Bookmark from "./bookmark";
-import QualitiesList from "./qualitiesList";
-import Table from "./table";
+import Table from "../common/table";
+import Bookmark from "../common/bookmark";
+import Qualities from "./qualities";
 import { Link } from "react-router-dom";
 
 const UsersTable = ({
@@ -24,7 +22,7 @@ const UsersTable = ({
         },
         qualities: {
             name: "Качества",
-            component: (user) => <QualitiesList qualities={user.qualities}/>
+            component: (user) => <Qualities qualities={user.qualities}/>
         },
         professions: {
             path: "profession.name",
@@ -60,23 +58,12 @@ const UsersTable = ({
         }
     };
 
-    return (
-        //first method
-        // <Table onSort={onSort} selectedSort={selectedSort} columns={columns} data={users}/>
-
-        // second method with children
-        <Table>
-            <TableHeader {...{
-                onSort,
-                selectedSort,
-                columns
-            }} />
-            <TableBody {...{
-                columns,
-                data: users
-            }} />
-        </Table>
-    );
+    return <Table
+        onSort={onSort}
+        selectedSort={selectedSort}
+        columns={columns}
+        data={users}
+    />;
 };
 
 UsersTable.propTypes = {
