+++ src/app/layouts/login.jsx
@@ -1,102 +1,7 @@
-import React, { useEffect, useState } from "react";
-import TextField from "../components/textField";
-import { validator } from "../utils/validator";
+import React from "react";
+import LoginForm from "../components/ui/loginForm";
 
 const Login = () => {
-    const [data, setData] = useState({
-        email: "",
-        password: ""
-    });
-    const [errors, setErrors] = useState({});
-
-    const handleChange = ({ target }) => {
-        setData((prevState) => ({
-            ...prevState,
-            [target.name]: target.value
-        }));
-    };
-
-    const validatorConfig = {
-        email: {
-            isRequired: {
-                message: "Электронная почта обязательна для заполнения"
-            },
-            isEmail: {
-                message: "Email введен не корректно"
-            }
-        },
-        password: {
-            isRequired: {
-                message: "Пароль обязателен для заполнения"
-            },
-            isCapitalSymbol: {
-                message:
-                    "Пароль должен содержать как минимум одну заглавную букву"
-            },
-            isContainDigit: {
-                message: "Пароль должен содержать как минимум одну цифру"
-            },
-            min: {
-                message: "Пароль должен содержать минимум 8 символов",
-                value: 8
-            }
-        }
-    };
-
-    useEffect(() => {
-        validate();
-    }, [data]);
-
-    const validate = () => {
-        const errors = validator(data, validatorConfig);
-        // for (const fieldName in data) {
-        //     if (data[fieldName].trim() === "") {
-        //         errors[fieldName] = `${fieldName} необходимо заполнить поле`;
-        //     }
-        // }
-        setErrors(errors);
-        return Object.keys(errors).length === 0;
-    };
-
-    const isValid = Object.keys(errors).length === 0;
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        const isValid = validate();
-        if (!isValid) return;
-        console.log(e);
-    };
-    return (
-        <div className="container mt-5">
-            <div className="row">
-                <div className=".col-md-6 .offset-md-3 shadow p-4">
-                    <h3 className="mb-4">Login</h3>
-                    <form onSubmit={handleSubmit}>
-                        <TextField
-                            label="Электронная почта"
-                            name="email"
-                            value={data.email}
-                            onChange={handleChange}
-                            error={errors.email}
-                        />
-                        <TextField
-                            label="Пароль"
-                            type="password"
-                            name="password"
-                            value={data.password}
-                            onChange={handleChange}
-                            error={errors.password}
-                        />
-                        <button
-                            type="submit"
-                            disabled={!isValid}
-                            className="btn btn-primary w-100 mx-auto"
-                        >
-                            Submit
-                        </button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    );
+    return <LoginForm/>;
 };
 export default Login;
