+++ src/app/layouts/users.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import UserPage from "../components/userPage";
-import UsersListPage from "../components/usersListPage";
+import UserPage from "../components/page/userPage";
+import UsersListPage from "../components/page/usersListPage";
 
 const Users = () => {
     const { userId } = useParams();
